+++ monitor/src/test/java/clazzfish/monitor/internal/ClassDiagnosticTest.java
@@ -18,10 +18,12 @@
 package clazzfish.monitor.internal;
 
 import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Collection;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.fail;
 
 /**
  * Unit-Test for {@link ClassDiagnostic}.
@@ -31,22 +33,36 @@ import static org.junit.jupiter.api.Assertions.*;
  */
 class ClassDiagnosticTest {
 
+    private static final Logger log = LoggerFactory.getLogger(ClassDiagnosticTest.class);
+
     @Test
     void getLoadedClassesFromGC() {
         Collection<Class<?>> loadedClasses = ClassDiagnostic.getLoadedClassesFromGC();
-        assertFalse(loadedClasses.isEmpty());
+        checkClasses(loadedClasses);
+    }
+
+    @Test
+    void getLoadedClassesFromVmClassHierarchy() {
+        Collection<Class<?>> loadedClasses = ClassDiagnostic.getLoadedClassesFromVmClassHierarchy();
+        checkClasses(loadedClasses);
     }
 
     @Test
     void getLoadedClassesFromStacktrace() {
         Collection<Class<?>> loadedClasses = ClassDiagnostic.getLoadedClassesFromStacktrace();
-        assertContains(loadedClasses, getClass());
+        checkClasses(loadedClasses);
     }
 
     @Test
     void getLoadedClasses() {
         Collection<Class<?>> loadedClasses = ClassDiagnostic.getLoadedClasses();
+        checkClasses(loadedClasses);
+    }
+
+    void checkClasses(Collection<Class<?>> loadedClasses) {
+        log.info("Checking {} loaded classes...", loadedClasses.size());
         assertContains(loadedClasses, getClass());
+        assertContains(loadedClasses, Object.class);
     }
 
     private void assertContains(Collection<Class<?>> classes, Class<?> expectedClass) {
@@ -57,5 +73,28 @@ class ClassDiagnosticTest {
         }
         fail("Class '" + expectedClass + "' not in " + classes);
     }
+
+//    @Test
+//    void getSystemClasses() throws IOException {
+//        Set<ModuleReference> references = ModuleFinder.ofSystem().findAll();
+//        assertFalse(references.isEmpty());
+//        references.forEach(System.out::println);
+//        for (ModuleReference ref : references) {
+//            try (ModuleReader moduleReader = ref.open()) {
+//                moduleReader.list().forEach(System.out::println);
+//            }
+//        }
+//    }
+//
+//    @Test
+//    void getVmLog() throws JMException {
+//        String mbeanName = "com.sun.management:type=DiagnosticCommand";
+//        Object result = ManagementFactory.getPlatformMBeanServer().invoke(
+//                new ObjectName(mbeanName),
+//                "vmClassHierarchy",
+//                new Object[]{new String[]{}},
+//                new String[]{"[Ljava.lang.String;"});
+//        System.out.println(result.toString());
+//    }
 
 }